Atualização master

+++ front-end/public/scripts/membros.js
@@ -1,8 +1,11 @@
 import membersCards from "./modules/membersCards.js"
 import urlParser from "./modules/urlParser.js"
+import Fork from "./modules/Fork.js"
+import Optional from "./modules/Optional.js"
 import deleteAnInvitation from "./requests/deleteAnInvitation.js";
 import gtHeaders from "./requests/gtHeaders.js";
 import postNewMember from "./requests/postNewMember.js";
+import swal from 'sweetalert';
 
 const mmbc = membersCards();
 const urlp = urlParser();
@@ -18,6 +21,7 @@ const templateInvitedCard = document.querySelector("template#t_membro_busca");
 const urlParams = urlp.mapVariables(location.href);
 
 searchButton.addEventListener('click', function getUserAndTravelToMakeAnInvitation(){
+        inviteBlock.innerHTML = "";
 
         const urls = { "searchUser": `http://localhost:3333/usuarios/buscar?q=${searchInput.value}`,
                        "getTravel": `http://localhost:3333/viagens/ler/${urlParams.travel_id}` }
@@ -33,7 +37,7 @@ searchButton.addEventListener('click', function getUserAndTravelToMakeAnInvitati
                .then(values => ({"usuario": values[0], "viagem": values[1]}))
                .then(json => mmbc.buildToInviteCard(templateInvitedCard, json))
                .then(card => inviteBlock.appendChild(card))
-               .then(e => console.log(e))
+               .then(e => {console.log(e)})
 })
 
 window.addEventListener('load', function getActiveTravels(){
@@ -45,20 +49,58 @@ window.addEventListener('load', function getActiveTravels(){
 
         fetch(urlToGetUserMembersOfTravel, init)
                 .then(res => res.json())
-                .then(json => Array.isArray(json)? json: false)
-                .then(json => json? json.reduce((acc, current) => current.aceito? (acc.aceitos.push(current), acc): (acc.pendentes.push(current), acc), {"aceitos":[], "pendentes":[]}): false)
-                .then(embarques => embarques? (embarques.aceitos.map(data => mmbc.buildMemberCard(templateMemberCard, data)).forEach(card => membersBlock.appendChild(card)), embarques): (membersBlock.innerText = "Nenhum membro", false))
-                .then(embarques => embarques? embarques.pendentes.map(data => mmbc.buildGuestCard(templateGuestCard, data)).forEach(card => guestsBlock.appendChild(card)): guestsBlock.innerText = "Nenhum convite")
+                .then(json => {
+                        const fork = Fork.of(json)
+                                         .between("aceitos", "pendentes")
+                                         .flatReduce(curr => curr.aceito)
+ 
+                        const toAppendMembersCards = Optional.of(fork.aceitos)
+                                                             .flatMap(data => mmbc.buildMemberCard(templateMemberCard, data))
+                                                             .flatMap(card => membersBlock.appendChild(card))
+                                                             .getOrElse(() => membersBlock.innerText = "Nenhum membro")
+ 
+                        const toAppendGuestsCards = Optional.of(fork.pendentes)
+                                                            .flatMap(data => mmbc.buildGuestCard(templateGuestCard, data))
+                                                            .flatMap(card => guestsBlock.appendChild(card))
+                                                            .getOrElse(() => guestsBlock.innerText = "Nenhum convite")
+                })
                 .catch(e => console.log(e))
 })
 
 window.addEventListener("guestCardCancelButtonClick", function cancelInvitation(e){
+        swal({
+                title: "Deseja cancelar este convite?",
+                text: "Você pode realizar esse convite novamente...",
+                icon: "warning",
+                buttons: true,
+                dangerMode: true,
+              })
+              .then((willDelete) => {
+                if (willDelete) {
+                        const request = deleteAnInvitation(gtHeaders.authorized(), e.detail)
 
-        const request = deleteAnInvitation(gtHeaders.authorized(), e.detail)
-
-        fetch(request.url, request.init)
-                .then(res => res.json())
-                .then(res => console.log(res.message))
+                        fetch(request.url, request.init)
+                                .then(res => res.json())
+                                .then(res => {
+                                        if(res.message !== undefined){
+                                                swal (`${res.message}` , { 
+                                                        icon: "error",
+                                                        buttons : false, 
+                                                        timer : 2000 })
+                                                .then((value) => window.location.href = `membros.html?travel_id=${urlParams.travel_id}`);
+                                        }
+                                })
+                                
+                        swal("Convite Deletado!", {
+                        icon: "success",
+                        buttons : false,
+                        timer : 2000
+                        })
+                        .then((value) => window.location.href = `membros.html?travel_id=${urlParams.travel_id}`);
+                } else {
+                        swal("Seu convite continua aguardando aprovação!")
+                }
+              })      
 })
 
 window.addEventListener("guestCardInviteButtonClick", function inviteToTravel(e){
@@ -67,5 +109,19 @@ window.addEventListener("guestCardInviteButtonClick", function inviteToTravel(e)
 
         fetch(request.url, request.init)
                 .then(res => res.json())
-                .then(res => console.log(res.message))
+                .then(res => {
+                        if(res.message !== undefined){
+                                swal (`${res.message}` , { 
+                                        icon: "error",
+                                        buttons : false, 
+                                        timer : 2000 })
+                                .then((value) => window.location.href = `membros.html?travel_id=${urlParams.travel_id}`);
+                        } else {
+                                swal ("Convite Realizado!" , { 
+                                        icon: "success",
+                                        buttons : false, 
+                                        timer : 2000 })
+                                .then((value) => window.location.href = `membros.html?travel_id=${urlParams.travel_id}`);
+                        }
+                })
 })
